Fix fake driver handling of failure replies
The driver reply() method is actually passed a full sys.exc_info()
tuple.

This was masked in the unit tests because the driver ended up basically
doing:

  raise (ValueError, ValueError, ...)

which caused a new ValueError to be instantiated and the test was
satisified. However, if an exception type has some required arguments,
you'll get a TypeError when this statement attempts to instantiate it
with no arguments.

Change-Id: I4af9c5084954d7b9c5f02cdae3387d17c206985b