Fix rabbit driver use of ssl.PROTOCOL_SSLv3
The rabbit driver required that ssl.PROTOCOL_SSLv3 is defined, but
not all Python implementations define it.

The code was changed so that the "sslv3" value is only allowed for
the "kombu_ssl_version" version if ssl.PROTOCOL_SSLv3 is defined in
the Python standard library.

The title of the commit in master was:

 Remove the use of PROTOCOL_SSLv3

but this was changed since it didn't describe the actual change.

Closes-Bug: #1395095
Change-Id: I2c1977c3bfc1923bcb03744e909f2e70c7fdb14c
(cherry picked from commit 42f55a1dda96d4ceecf8cca5fba9cd723673f6e3)