Add a new aioeventlet executor
The new executor supports trollius coroutines, explicit asynchronous
programming, in addition to eventlet greenthreads, implicit asynchronous
programming.

The new AsyncioEventletExecutor class is based on the EventletExecutor
class and so it is compatible with it. The aioeventlet executor can be
used to replace the eventlet executor, but it requires an aioeventlet
event loop running in the thread running the executor (usually the main
thread). See AsyncioEventletExecutor docstring for an example how to
setup such event loop.

The aioeventlet module implements the asyncio API (PEP 3156) on top of
eventlet, see aioeventlet documentation:
http://aioeventlet.readthedocs.org/

The change adds an unit test with an endpoint implemented as a trollius
coroutine.

The executor is not supported on Python 3 yet because of eventlet issues
with monkey patching.

Implements: blueprint greenio-executor
Change-Id: I7a78ed998719a703077232726f66d882463b1297