kombu: fix driver loading with kombu+qpid scheme
When a url looks like: kombu+qpid:///host:port/ the driver fail
to load.

This change fixes that, adds a warning message that this kind of
URL is experimental and not yet supported.

Also our Consumer code use internal kombu API that can be optionnal
implemented by the kombu transport (message_to_python),
so check that this one exists before using it.

In the future, we should use kombu.messaging.Consumer/Publisher
instead of Consumer/Publisher implementation to avoid such hack...

Change-Id: I066d57c23bff922c5734ab036b6ca8e1608e5c6a