Make sure zmq can work with redis
In ZmqDriver's listen method, it calls create_consumer
three times. After the first call, the keys related to this
topic in redis reads like this:
  "topic": set(["topic.host"])
  "topic.host": ""

If the second call, it tries to add the following keys:
  "topic.host": set(["topic.host.host"])
  "topic.host.host": ""

But the key "topic.host" already exists as a string type.
So the error occurs.

To resolve this problem, change the value of keys from string
to set.

Change-Id: Ic801393d492d2656fcfd8b87f1d2efc6ab3bbd62
Closes-Bug: #1290772