Port zmq driver to Python 3
With eventlet 0.16, it becomes possible to run Oslo Messaging tests on
Python 3 with eventlet.

This change ports the zmq driver to Python 3:

* encode the topic explicitly to UTF-8
* use a list comprehension instead of map() to also get a list
  on Python 3 (not a generator)

The following eventlet change is needed to run tests:
https://github.com/eventlet/eventlet/pull/187

Related eventlet issue:
https://github.com/eventlet/eventlet/issues/185

I will propose a different change to enable tests with eventlet enabled
when a release of eventlet including this fix will be available.

Change-Id: Ic8fec515cfa757e08ffb9604e3bfb2e87d08f3d8