Fix the bug redis do not delete the expired keys
When using redis as zmq backend,
the keys in redis maybe like follows:

 "service"
 "service.host1"
 "service.host1.host1"
 "service.host2"
 "service.host2.host2"
 "service.host3"
 "service.host3.host3"

If we stopped the service on host1, the key named
service.host1 and service.host1.host1 shoule be
delete. As the result, the topicexchange message
would not sent to host1. But the fact is
service.host1 still exsit, and topicexchange
message may still sent to host1.

To resolve this problem , change the comparison of
ttl function from -1 to -2 according to
http://redis.io/commands/ttl

Change-Id: I5c0af97019fffed6e949eb58d7d60c85f5b08ea1
Closes-Bug:#1417464
(cherry picked from commit 1958f6e549d3769202960f5af93982afea74cd80)