rabbit: Set timeout on the underlying socket
--
NOTE(mriedem): This is two commits squashed to address a problem with
the rabbitmq heartbeat patch in stable/kilo (since oslo.messaging 1.8.1).
--

rabbit: Remove unused stuffs from publisher

The publisher code is over engineered, it allows to override
everything, but this is never used.

None of the child Class have the same signature, sometimes
the constructor use the parameter name as the parent class but for
a different purpose, that make the code hard to read.

It's was never clear which options is passed to the queue and the
exchange at this end to kombu.

This changes removes all of that stuffs, and only use the kombu
terminology for publisher parameters.

(cherry picked from commit cca84f66d49114ce4ae85af4bbf03a14bda79121)

--------------------------------------------

rabbit: Set timeout on the underlying socket

They are some case where the underlying can be stuck
until the system socket timeout is reached, but in oslo.messaging
we very often known that is not needed to wait for ever because
the upper layer (usualy the application) expect to return after
a certain period.

So this change set the timeout on the underlying socket each we can
determine that is not needed to wait more.

Closes-bug: #1436788
Change-Id: Ie71ab8147c56eaf672585da107bec8b22af9da6c
(cherry picked from commit 77f952a1f7d87c35fb5d7eb7e052c5b9def59b22)