rabbit: redeclare consumers when ack/requeue fail
In case the acknowledgement or requeue of a message fail,
the kombu transport can be disconnected

In this case, we must redeclare our consumers.

This changes fixes that.

This have no tests because the kombu memory transport we use in our tests
cannot be in disconnected state.

Closes-bug: #1448650

(cherry picked from commit 415db68b67368d7c8aa550e7108122200816e665)

Conflict is due to refactoring on master branch.

Conflicts:
    oslo_messaging/_drivers/impl_rabbit.py

Change-Id: I5991a4cf827411bc27c857561d97461212a17f40