rabbit: fixes a logging issue
When uses the default port, kombu.Connection.port is None

So we replace the usage of kombu.Connection.port per
kombu.Connection.info(), to get the default value.

Also some transport driver have 'None' for hostname or port as default
value, so replace usage of '%d' per '%s', to ensure the logging never fail.
Event the output of the log is less sexy.

Change-Id: I89ca1982246146717015253bd4cc26f992381584
Closes-bug: #1452189