Use the warn_eventlet_not_patched util function
This code can share the check for eventlet and warn
when the thread module is not patched (which will be
a problem).

When it is not patched the following warning gets
emitted:

"RuntimeWarning: It is highly recommended that when eventlet"
" is used that the ['thread'] modules are monkey patched when"
" using the 'oslo.messaging eventlet executor' (to avoid spurious"
" or unexpected lock-ups and/or hangs)"

This will not be needed once a new futurist release happens
but until then standardize on using that util library to emit
these warnings.

Change-Id: I7277fe19a73d8094a0acf033c75dc881e0d1e533