Fix AMQP 1.0 functional and unit test failures
As a result of upgrading proton to 0.10, some behavioral changes have
triggerd failures in the gate.

This patch fixes a subprocess call and syntax error in the AMQP1.0 unit tests.

This patch also fixes the functional test setup script to correctly
configure olter versions of QPIDD.

Both these fixes are necessary to get the gate to pass.

Change-Id: I5ca8476999fc953688c76c302a59bedf70658c8b
Closes-Bug: #1494794
Closes-Bug: #1494918