rabbit: shuffle hosts before building kombu URL
This restores behavior as it was back in the Juno cycle.  This was
changed in a refactor during the Kilo cycle (973301aa) when the kombu
failover_strategy was set to shuffle.  But the failover_strategy only
affects failovers; it does not influence which host receives the
initial connection.

Thus, if multiple hosts are given, the first host will receive all of
the initial connections.  By shuffling the hosts before building the
URL given to kombu, rudimentary load balancing is achieved and
connections are more evenly spread across broker nodes.

Change-Id: Ieecaf5f286f49302db29df9854359ef5d36b7a89
(cherry picked from commit f2beb5eaf158f76cdb8be0ca72965503be423ee2)