Implements more smart retrying
This patch makes difference between messages which are sent
with retry=0 and retry!=0. In first case we consider that
request could be not idempotent and we should guarantee
that such messages we be not sent more then one time.
Therefore we use consuming without acknowledgements for such
messages and consuming with acknowledgements for messages which
could be processed twice to avoid message loses

Change-Id: I7dc1705e84dc6a7671bf1f379d4ef4980e41491e