Use round robin failover strategy for Kombu driver
Shuffle strategy we use right now leads to increased reconnection time
and provides no benefit. Sometimes it might lead to RPC operations
timeout because the strategy provides no guarantee on how long the
reconnection process will take. See the referenced bug for details.

On the other side, round-robin strategy provides least achievable
reconnection time. It also provides guarantee that if K of N RabbitMQ
hosts are alive, it will take at most N - K + 1 attempts to
successfully reconnect to RabbitMQ cluster.

With shuffle strategy during failover clients connect to random hosts
and so the load is distributed evenly between alive RabbitMQs.
But since we shuffle list of hosts before providing it to Kombu, load
will be distributed evenly with round-robin strategy as well.

DocImpact
A new configuration option kombu_failover_strategy for Kombu driver is
added. It determines how the next RabbitMQ node is chosen in case the
one we are currently connected to becomes unavailable. It takes effect
only if more than one RabbitMQ node is provided in config. Available
options are:

 * round-robin: each RabbitMQ host in the list is tried in cycle until
   oslo.messaging successfully connects. Since oslo.messaging
   shuffles list of RabbitMQ hosts, the order of hosts in the cycle
   will be random and will not depend on order provided in config.

 * shuffle: oslo.messaging selects a random host from the list and
   tries to connect to it. If connection fails, oslo.messaging repeats
   attempt to connect to another random host. Oslo.messaging stops
   once it successfully connects to a host. Note that in each
   iteration a host to connect is selected independently of previous
   iterations, i.e. it might happen that oslo.messaging will try to
   connect to the same host several times in a row.

The option's default value is round-robin. Before the option was
introduced, the default strategy was shuffle. For the reasoning,
see the main body of the commit message and the referenced bug.

Closes-Bug: #1519851
Change-Id: I9a510c86bd5a6ce8b707734385af1a83de82804e