.testr.conf: revert workaround of testtools bug
Revert the change I6507e693fc929e03884f933bbda241f744d3a7c0. The
testtools bug was fixed, the "| cat" workaround is no more needed.

Fix for subunit (for testtools), workaround the eventlet bug:
https://github.com/testing-cabal/subunit/pull/14

Fix of the root cause in eventlet:
https://github.com/eventlet/eventlet/issues/248

The bug was limited to Python 3 and related to non-blocking stdout.

Change-Id: I207b7a7e82cbd7e5848231115b3577be7b600638
Related-Bug: 1492505