Implements pika thread safe connection
ThreadSafePikaConnection works over SelectConnection which has
asynchronous interface and provides synchronous interface
It allows to use single connection in concurrent environment
with nonblocking io approach. For this goal internal thread is added
for processing ioloop.

Also this approach allows to remove poller's thread and use ioloop's thread
for calling on_incoming calbacks. It is also done by this patch

This patch is important to use single connection for whole process in future

Change-Id: I12c16715f6bf8a99e438bc054f9d0132a09cecf3
Depends-On: I41a768c5624fa2212257ce20bf9a67d09de0c4ab