Log the unique_id in listener than msg_id
When use notify functionality of oslo.message, the message listener
(message receiver) will log the msg_id, but the msg_id is only generated
when send a RPC CALL message. So that many useless message like
"received message msg_id: None reply to None" will be recorded in log
file. This change will record the unique_id which is useful for tracing
the messages from sender to receiver.

Change-Id: I61a174b2af60e0e7fb61436d6e20d01b2c413c45