Use single producer and to avoid an exchange redeclaration
Current implementation of the Connection class on each call of _publish method
creates its own channel-bounded producer and if auto_declared
is set to True then will be send exchange declaration request to the broker
despite the fact that the exchange might be created by previous calls.

It leads to the significant ~11-30%  overhead.
I got the following results with notification client & server in the simulator:
 - http://s32.postimg.org/jzgkbs1qt/notify_client.png

I suggest to retain the sets of already declared exchanges and queues per each
connection instance and to use a single channel-bounded producer. No needs to
create a producer on each call, it is just wrapper over channel.

Change-Id: I89564dd1aa079476c68c50979b6b79f8889db5d9