[zmq] Don't skip non-direct message types
If using the router proxy we need to be able to
dispatch all types of messages over the routers,
not limiting them to direct types only.

Also added fanout possibility to do on a client-side,
so we can reduce latency on a proxy, and not using
pub-sub for fanout when we don't want it.

Change-Id: Ic88c306c1d386dd582cbccfc5719fba5668a9db8