Make TransportURL.parse aware of transport_url
For no real reason, the transport_url confi option was not read by the
TransportURL class while other config options are read directly from
configuration. Instead the transport_url was always passed to
TransportURL.parse().

This change allows url to be optional for TransportURL.parse() and read
the url from transport_url config when not provided.

Change-Id: Idfe9aa37bf3a409d2c1f76d99ea3d2bdf9837193