Avoid sending cast after server shutdown in functional test
The server test fixtures attempt to send a cast to the server during
server cleanup.  This cast is sent after the server is shut down, so
the cast only goes as far as the message bus.  This is fine for
intermediary-based message buses such as brokers, but for
point-to-point connections this can cause the cast to hang since there
is no recipient for the cast.

Change-Id: I2967ef21deb12dc1f84f6dc97a894778aeb42666
Closes-Bug: #1616487