Flesh out transport_url help
This is a complicated config opt and the single line help text
doesn't provide anywhere near enough detail for a user to properly
configure it.  This change adds the expected format from the
TransportURL docstring and leaves a link to the docs for that
class since the full details are probably too verbose to put in the
help text in their entirety.

Change-Id: I66122bf9c619d35bcb3caadfde9c0dcdb5cd9095