Revert "rabbit: Don't prefetch when batch_size is set"
This reverts commit c881baed29db49c5710795496cb07907e35ceaba.

Eventhough batch_size is set, rabbitMQ prefetch count is not changed
at the first time of connection creation. It works only connection reset
cases(i.e. rabbitMQ restart). So this patch can not fix original issue
  https://bugs.launchpad.net/ceilometer/+bug/1551667

Moreover, it makes another bug that messages of sameple queues are not
consumed properly. So revert it.

Change-Id: Ia8ebee8e2a670e46b6a68859bc30e717bd56ed7e
Signed-off-by: Wonil Choi <wonil22.choi@samsung.com>
Closes-bug: 1759755