Correct usage of transport_url in example
The example usage of transport_url in ConfFixture was not corrected when
the rest of the code was switched to using 'fake:/' for transport_url.
Since other projects will refer to this example to review changes, it
should be correct.

Also change the transport URLs used in test_transport to correspond to
the correct way of specifying virtual hosts.

Change-Id: If8e0a869e02fda2d8d2c86d82ef83ffc3a4361e4