import zuul job settings from project-config
This is a mechanically generated patch to complete step 1 of moving
the zuul job settings out of project-config and into each project
repository.

Because there will be a separate patch on each branch, the branch
specifiers for branch-specific jobs have been removed.

The older documentation job template is used because the new PTI job
does not work with this old branch of this repository.

The python 3.5 job is dropped because it is failing and we do not need
to keep testing it on the old branch.

See the python3-first goal document for details:
https://governance.openstack.org/tc/goals/stein/python3-first.html

Depends-On: https://review.openstack.org/592029
Change-Id: I796c720d0ca6ecb104017a3b45167eca8fd3dd30
Story: #2003250